+++ tests/Main.hs
@@ -24,20 +24,22 @@ main = hspec spec
 spec :: Spec
 spec = do
   describe "levenshtein" $ do
-    testSwap levenshtein_
+    testSwap levenshtein
     context "with concrete examples" $ do
-      testPair levenshtein_ "kitten"   "sitting" 3
-      testPair levenshtein_ "cake"     "drake"   2
-      testPair levenshtein_ "saturday" "sunday"  3
-      testPair levenshtein_ "red"      "wax"     3
-      testPair levenshtein_ "lucky"    "lucky"   0
-      testPair levenshtein_ ""         ""        0
+      testPair levenshtein "kitten"   "sitting" 3
+      testPair levenshtein "cake"     "drake"   2
+      testPair levenshtein "saturday" "sunday"  3
+      testPair levenshtein "red"      "wax"     3
+      testPair levenshtein "a😀c"     "abc"     1
+      testPair levenshtein "lucky"    "lucky"   0
+      testPair levenshtein ""         ""        0
   describe "levenshteinNorm" $ do
     testSwap levenshteinNorm
     testPair levenshteinNorm "kitten"   "sitting" (4 % 7)
     testPair levenshteinNorm "cake"     "drake"   (3 % 5)
     testPair levenshteinNorm "saturday" "sunday"  (5 % 8)
     testPair levenshteinNorm "red"      "wax"     (0 % 1)
+    testPair levenshteinNorm "a😀c"     "abc"     (2 % 3)
     testPair levenshteinNorm "lucky"    "lucky"   (1 % 1)
     testPair levenshteinNorm ""         ""        (1 % 1)
   describe "damerauLevenshtein" $ do
@@ -47,6 +49,7 @@ spec = do
     testPair damerauLevenshtein "nose"          "ones"      2
     testPair damerauLevenshtein "thing"         "sign"      3
     testPair damerauLevenshtein "red"           "wax"       3
+    testPair damerauLevenshtein "a😀c"          "abc"       1
     testPair damerauLevenshtein "lucky"         "lucky"     0
     testPair damerauLevenshtein ""              ""          0
   describe "damerauLevenshteinNorm" $ do
@@ -56,6 +59,7 @@ spec = do
     testPair damerauLevenshteinNorm "nose"          "ones"      (1 % 2)
     testPair damerauLevenshteinNorm "thing"         "sign"      (2 % 5)
     testPair damerauLevenshteinNorm "red"           "wax"       (0 % 1)
+    testPair damerauLevenshteinNorm "a😀c"          "abc"       (2 % 3)
     testPair damerauLevenshteinNorm "lucky"         "lucky"     (1 % 1)
     testPair damerauLevenshteinNorm ""              ""          (1 % 1)
   describe "hamming" $ do
@@ -66,7 +70,7 @@ spec = do
     testPair hamming "2173896" "2233796" (Just 3)
     testPair hamming "toned"   "roses"   (Just 3)
     testPair hamming "red"     "wax"     (Just 3)
-    testPair hamming "a😀c"    "abc"     (Just 1)
+    testPair hamming "a😀c"    "abc"      (Just 1)
     testPair hamming "lucky"   "lucky"   (Just 0)
     testPair hamming ""        ""        (Just 0)
     testPair hamming "small"   "big"     Nothing
@@ -86,6 +90,7 @@ spec = do
     testPair jaro "five"   "ten"      (0  % 1)
     testPair jaro "ten"    "five"     (0  % 1)
     testPair jaro "lucky"  "lucky"    (1  % 1)
+    testPair jaro "a😀c"   "abc"      (7  % 9)
     testPair jaro ""       ""         (0  % 1)
   describe "jaroWinkler" $ do
     testPair jaroWinkler "aa" "a"            (17 % 20)
@@ -103,6 +108,7 @@ spec = do
     testPair jaroWinkler "five"   "ten"      (0  % 1)
     testPair jaroWinkler "ten"    "five"     (0  % 1)
     testPair jaroWinkler "lucky"  "lucky"    (1  % 1)
+    testPair jaroWinkler "a😀c"   "abc"      (4  % 5)
     testPair jaroWinkler ""       ""         (0  % 1)
 
 -- | Test that given function returns the same results when order of