+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Text Metrics 0.3.0
+
+* All functions are now implemented in pure Haskell.
+
+* All functions return `Int` or `Ratio Int` instead of `Natural` and `Ratio
+  Natural`.
+
 ## Text Metrics 0.2.0
 
 * Made the `levenshtein`, `levenshteinNorm`, `damerauLevenshtein`, and